Завершение. Создание хука и корректировки
